Standardize dependencies and imports
This commit makes some reasonable assumptions about the use cases of pymzML, and uses those
assumptions to significantly clean up the code, simplify it, and make it more flexible.

The assumptions are:
1. The library should only run on python >= 3.4
2. Numpy is no longer optional

The first is explicitly stated by the project, and the second is an easy criterion: no one doing
calculation on mass spec should _not_ have numpy installed; especialyl if they are using python3,
it is virtually impossible to have a python system that does not have a successful version of numpy.

In addition to the above assumptions, the following flexibility improvements are implemented.
3. Plotly should not be made a default requirement: it often requires many other packages, which
   systems may not want to install, especially if they are using pymzml on a server
4. If pynumpress is in fact optional, as stated in the docs, it should not be required when
   importing pymzml

Based on the assumptions and flexibility improvements above, this commit:
1. Makes python 3.4 a requirement in setup.py
2. Makes numpy a non-optional dependency in setup.py
3. Moves plotly to an 'extras' requirement, installed with by `pip install pymzml['plot']`
4. Moves pynumpress to an 'extras' requirement
5. Adds a 'full' exstra that installs both
6. standardizes setup.py and requirements.txt to agree with one another in version numbers
7. Fails gracefully by warning a user when they are missing an optional dependency, and only
   terminates execution when code that uses that dependency is called
8. Updates imports to use python3 importing rules and consistently import packages across the
   project. avoiding circular dependencies and removing needless code
9. Removes obsolete code that implemented functionality for users without numpy
10. Removes legacy Python 2-specific code, such as `from __future__ import print_function`
11. Updates the pynumpress dependency to 0.0.4, the latest release

For more info on extras, see:
https://setuptools.readthedocs.io/en/latest/setuptools.html#declaring-extras-optional-features-with-their-own-dependencies